added struct rusage, defined wait2(), added entry for wait2()

+++ kernel/pstat.h
@@ -8,5 +8,5 @@ struct pstat {
 
 
 struct rusage {
-  uint64 cputime;
+  uint cputime;
 };
